fixed dashboard private route bug
