+++ src/components/layout/DashboardLayout/DashboardLayout.tsx
@@ -5,32 +5,38 @@ import styles from "./dashboardLayout.module.scss";
 import { useNavigate } from "react-router-dom";
 import { AuthContext } from "../../../context/Auth";
 import { AuthContextType } from "../../../interfaces/auth";
+import { getFromLocalStorage } from "../../../utils/localStorage";
 
 
 const DashboardLayout = ({ children }: LayoutProps) => {
 
-    const { isLoggedIn } = useContext(AuthContext) as AuthContextType
+    const { logout } = useContext(AuthContext) as AuthContextType
     const navigate = useNavigate();
+    const logoutInterval = 120
 
     useEffect(() => {
+        const isLoggedIn = getFromLocalStorage("isAuthenticated")
         if (!isLoggedIn) {
             navigate("/")
         }
-    }, [isLoggedIn, navigate])
+    }, [navigate])
+
+    useEffect(() => {
+        const interval = setInterval(() => {
+            logout();
+            navigate("/")
+        }, (logoutInterval * 1000))
+
+        return () => clearInterval(interval);
+    }, [logout, navigate])
 
     return (
-        <>
-            {
-                isLoggedIn ? (
-                    <div>
-                        <div className={styles.nav}><div className={styles.logo}><Logo /></div></div>
-                        <div className={styles.holder}>
-                            {children}
-                        </div>
-                    </div>
-                ) : null
-            }
-        </>
+        <div>
+            <div className={styles.nav}><div className={styles.logo}><Logo /></div></div>
+            <div className={styles.holder}>
+                {children}
+            </div>
+        </div>
     )
 }
