+++ src/interfaces/auth.tsx
@@ -1,7 +1,6 @@
 export type AuthContextType = {
     login: (email: string, password:string)=>string;
     logout: ()=>void;
-    isLoggedIn: boolean;
 }
 
 export type UserData = {
