+++ src/styles/_mixins.scss
@@ -46,3 +46,15 @@
     color: #E60A2B;
     font-size: $textsmall;
 }
+
+@mixin baseinput(){
+    border-radius: 0.375rem;
+    border: 1px solid $headertext;
+    outline: 0.375rem solid #F0F0F0;
+    @include block;
+    
+    &:focus {
+        outline: 0.25rem solid #F0F0F0;
+        border: 1px solid $brandgreen;
+    }
+}
